+++ src/backend/routers/simulation_mapper.py
@@ -6,28 +6,29 @@ from fastapi import APIRouter, Request, Response, Form, Query
 from fastapi.responses import RedirectResponse
 
 from services.implementions.auth_service import Auth_Service
+from services.implementions.timeline_service import Timeline_Service
 from services.implementions.user_service import User_Service
 
 router = APIRouter()
 
 auth_service = Auth_Service()
 user_service = User_Service()
-
+timeline_service = Timeline_Service()
 
 @router.get("/latest")
 def _(response: Response):
     response.status_code = 200
-    return {"latest": auth_service.get_latest()}
+    return {"latest": timeline_service.get_latest()}
 
 @router.get("/msgs")
 def _(latest: Union[str, None] = Query(default=-1)):
-    auth_service.record_latest(latest)
+    timeline_service.record_latest(latest)
     return RedirectResponse("/api/timelines/public", status_code=307)
 
 # This is a route that bypasses authorization and our session so it is implemented here
 @router.get("/msgs/{username}", status_code=204)
 def _(username: str, latest: Union[str, None] = Query(default=-1)):
-    auth_service.record_latest(latest)
+    timeline_service.record_latest(latest)
     return RedirectResponse(f"/api/timelines/{username}?PER_PAGE={latest}", status_code=307)
 
 class MessageBody(BaseModel):
@@ -35,12 +36,13 @@ class MessageBody(BaseModel):
 # This is a route that bypasses authorization and our session so it is implemented here
 @router.post("/msgs/{username}", status_code=200)
 def _(response: Response, username: str, body: MessageBody, latest: Union[str, None] = Query(default=-1)):
-    auth_service.record_latest(latest)
+
     user_id = user_service.get_user_id_from_username(username)
     if user_id is None:
-        response.status_code = 403
-        return {"error": "username already exists"}
+        response.status_code = 404
+        return {"error": "user doesn't exist"}
     
+    timeline_service.record_latest(latest)
     user_service.post_message(user_id, body.content)
     # passing a body when redirecting is not supported
     # https://github.com/tiangolo/fastapi/issues/3963
@@ -55,11 +57,11 @@ class Registration(BaseModel):
 
 @router.post("/register", status_code=204)
 def _(response: Response, body: Registration, latest: Union[int, None] = Query(default=-1)):
-    auth_service.record_latest(latest)
     if auth_service.check_if_user_exists(body.username):
         response.status_code = 403
         return {"error": "username already exists"}
     else:
+        timeline_service.record_latest(latest)
         response.status_code = 204
         auth_service.register_user(body.username, body.email, body.pwd)
         return {"success": "register success"}
@@ -72,22 +74,28 @@ class FollowMessage(BaseModel):
 @router.get("/fllws/{username}")
 def _(username: str, response: Response, no: Union[str, None] = Query(default=100), latest: Union[str, None] = Query(default=-1)):
     user_id = user_service.get_user_id_from_username(username)
+
     if not user_id:
         response.status_code = 404
         return {"error": "user doesn't exist"}
+    
+    timeline_service.record_latest(latest)
 
     followers = user_service.get_all_followers(user_id, no)
-    auth_service.record_latest(latest)
+
     response.status_code = 200
     follower_names = [f["username"] for f in followers]
     return {"follows": follower_names}
 
 @router.post("/fllws/{username}")
 def _(username: str, response: Response, body: FollowMessage, latest: Union[str, None] = Query(default=-1)):
-    auth_service.record_latest(latest)
     user_id = user_service.get_user_id_from_username(username)
+
     if user_id is None:
-        return Response(status_code=403)
+        response.status_code = 404
+        return {"error": "user doesn't exist"}
+    
+    timeline_service.record_latest(latest)
 
     if body.follow is not None:
