+++ src/backend/services/implementions/auth_service.py
@@ -1,5 +1,6 @@
 from services.interfaces.auth_service_interface import Auth_Service_Interface
 from repos.implementations.auth_repo import Auth_Repo
+import hashlib
 
 class Auth_Service(Auth_Service_Interface):
 
@@ -10,13 +11,9 @@ class Auth_Service(Auth_Service_Interface):
         return self.auth_repo.check_if_user_exists(username)
     
     def validate_user(self, username, password):
-        return self.auth_repo.validate_user(username,password)
+        hashed_pw = hashlib.md5(password.encode())
+        return self.auth_repo.validate_user(username, hashed_pw.hexdigest() )
 
     def register_user(self, username, email, password):
-        self.auth_repo.register_user(username,email,password)
-
-    def record_latest(self, latest):
-        self.auth_repo.record_latest(latest)
-        
-    def get_latest(self):
-        return self.auth_repo.get_latest()
+        hashed_pw = hashlib.md5(password.encode())
+        self.auth_repo.register_user(username,email,hashed_pw.hexdigest())
