+++ src/backend/services/interfaces/auth_service_interface.py
@@ -18,11 +18,3 @@ class Auth_Service_Interface:
     @abstractmethod
     def register_user(self, username, email, password):
         raise NotImplementedError
-    
-    @abstractmethod
-    def record_latest(self, latest):
-        raise NotImplementedError
-    
-    @abstractmethod
-    def get_latest(self):
-        raise NotImplementedError
